omicsdm-server/requirements-mini.txt
# boto3==1.17.88
boto3==1.35.78
celery==5.4.0
cryptography==44.0.0
docker==7.1.0
flask-restx==1.3.0
Flask==3.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.4
PyJWT==2.3.0
pytest==6.2.5
pytest-cov===3.0.0
flake8==4.0.1
flake8-html==0.4.1
redis==5.2.1
# requests==2.22.0
requests==2.32.3
Werkzeug==3.0.1
